requirements.txt
python-tk
python-pygame
vk_api